Move client roles methods to their own module
