+++ lib/clients.js
@@ -1,6 +1,7 @@
 'use strict';
 
 const privates = require('./private-map');
+const roles = require('./client-roles');
 const request = require('request');
 
 /**
@@ -13,9 +14,7 @@ module.exports = {
   update: update,
   remove: remove,
   getClientSecret: getClientSecret,
-  getRoles: getRoles,
-  getRole: getRole,
-  createRole: createRole
+  roles: roles
 };
 
 /**
@@ -231,132 +230,6 @@ function getClientSecret (client) {
         }
 
         return resolve(body);
-      });
-    });
-  };
-}
-
-/**
-  A function to get the roles of a client
-  @param {string} realmName - The name of the realm(not the realmID) to remove - ex: master,
-  @param {string} id - The id of the client(not the client-id) to remove
-  @returns {Promise} A promise that resolves.
-  @example
-  keycloakAdminClient(settings)
-    .then((client) => {
-      client.clients.getRoles(realmName, id)
-        .then((clientRoles) => {
-          console.log(clientRoles)
-      })
-    })
- */
-function getRoles (client) {
-  return function getRoles (realmName, id) {
-    return new Promise((resolve, reject) => {
-      const req = {
-        url: `${client.baseUrl}/admin/realms/${realmName}/clients/${id}/roles`,
-        auth: {
-          bearer: privates.get(client).accessToken
-        },
-        json: true
-      };
-
-      request(req, (err, resp, body) => {
-        if (err) {
-          return reject(err);
-        }
-
-        if (resp.statusCode !== 200) {
-          return reject(body);
-        }
-
-        return resolve(body);
-      });
-    });
-  };
-}
-
-/**
-  A function to get a single role of a client
-  @param {string} realmName - The name of the realm(not the realmID) to remove - ex: master,
-  @param {string} id - The id of the client(not the client-id) to remove
-  @param {string} roleName - The id of the client(not the client-id) to remove
-  @returns {Promise} A promise that resolves.
-  @example
-  keycloakAdminClient(settings)
-    .then((client) => {
-      client.clients.getRole(realmName, id, roleName)
-        .then((clientRole) => {
-          console.log(clientRole)
-      })
-    })
- */
-function getRole (client) {
-  return function getRole (realmName, id, roleName) {
-    return new Promise((resolve, reject) => {
-      const req = {
-        url: `${client.baseUrl}/admin/realms/${realmName}/clients/${id}/roles/${roleName}`,
-        auth: {
-          bearer: privates.get(client).accessToken
-        },
-        json: true
-      };
-
-      request(req, (err, resp, body) => {
-        if (err) {
-          return reject(err);
-        }
-
-        if (resp.statusCode !== 200) {
-          return reject(body);
-        }
-
-        // Since the create Endpoint returns an empty body, go get what we just imported.
-        return resolve(body);
-      });
-    });
-  };
-}
-
-/**
-  A function to create a new role.
-  @param {string} realmName - The name of the realm(not the realmID) - ex: master
-  @param {string} id - The id of the client(not the client-id) to remove
-  @param {object} newRole - The JSON representation of a role - http://www.keycloak.org/docs-api/3.0/rest-api/index.html#_rolerepresentation - name must be unique within the client
-  @returns {Promise} A promise that will resolve with the new clients object
-  @example
-  keycloakAdminClient(settings)
-    .then((client) => {
-      client.clients.createRole(realmName, id, newRole)
-        .then((createRole) => {
-        console.log(createRole) // [{...}]
-      })
-    })
- */
-function createRole (client) {
-  return function createRole (realm, id, newRole) {
-    return new Promise((resolve, reject) => {
-      const req = {
-        url: `${client.baseUrl}/admin/realms/${realm}/clients/${id}/roles`,
-        auth: {
-          bearer: privates.get(client).accessToken
-        },
-        body: newRole,
-        method: 'POST',
-        json: true
-      };
-
-      request(req, (err, resp, body) => {
-        if (err) {
-          return reject(err);
-        }
-
-        if (resp.statusCode !== 201) {
-          return reject(body);
-        }
-
-        // Since the create Endpoint returns an empty body, go get what we just imported.
-        return resolve(client.clients.getRole(realm, id, newRole.name));
       });
     });
   };
