+++ test/clients-test.js
@@ -297,7 +297,7 @@ test("Test getting a client's roles", (t) => {
     const realmName = 'master';
     const id = '379efc29-4b2e-403c-83b6-d9c9af43b24a'; // This is the master-realm client id from /scripts/kc-setup-for-tests.json
 
-    return client.clients.getRoles(realmName, id).then((roles) => {
+    return client.clients.roles.find(realmName, id).then((roles) => {
       t.equal(roles.length, 12, 'Should return 12 roles');
 
       const expectedRole = {
@@ -318,7 +318,7 @@ test("Test getting a client's roles - client id doesn't exist", (t) => {
   const id = 'not-a-real-id';
   const realmName = 'master';
   return kca.then((client) => {
-    return t.shouldFail(client.clients.getRoles(realmName, id), 'Could not find client', 'Should return an error that no user is found');
+    return t.shouldFail(client.clients.roles.find(realmName, id), 'Could not find client', 'Should return an error that no user is found');
   });
 });
 
@@ -331,7 +331,7 @@ test("Test getting a client's role", (t) => {
     const id = '379efc29-4b2e-403c-83b6-d9c9af43b24a'; // This is the master-realm client id from /scripts/kc-setup-for-tests.json
     const roleName = 'manage-identity-providers';
 
-    return client.clients.getRole(realmName, id, roleName).then((role) => {
+    return client.clients.roles.find(realmName, id, roleName).then((role) => {
       const expectedRole = {
         id: 'a16e820e-ae47-4ac9-82ba-683c0b866994',
         name: 'manage-identity-providers',
@@ -351,7 +351,7 @@ test("Test getting a client's role - client id doesn't exist", (t) => {
   const id = 'not-a-real-id';
   const roleName = 'not-a-real-role-name';
   return kca.then((client) => {
-    return t.shouldFail(client.clients.getRole(realmName, id, roleName), 'Could not find client', 'Should return an error that no client is found');
+    return t.shouldFail(client.clients.roles.find(realmName, id, roleName), 'Could not find client', 'Should return an error that no client is found');
   });
 });
 
@@ -362,7 +362,7 @@ test("Test getting a client's role - role name doesn't exist", (t) => {
   const id = '379efc29-4b2e-403c-83b6-d9c9af43b24a'; // This is the master-realm client id from /scripts/kc-setup-for-tests.json
   const roleName = 'not-a-real-role-name';
   return kca.then((client) => {
-    return t.shouldFail(client.clients.getRole(realmName, id, roleName), 'Could not find role', 'Should return an error that no role is found');
+    return t.shouldFail(client.clients.roles.find(realmName, id, roleName), 'Could not find role', 'Should return an error that no role is found');
   });
 });
 
@@ -377,7 +377,7 @@ test('Test create a client role', (t) => {
       description: 'A new role'
     };
 
-    return client.clients.createRole(realmName, id, newRole).then((addedRole) => {
+    return client.clients.roles.create(realmName, id, newRole).then((addedRole) => {
       t.equal(addedRole.name, newRole.name, `The name should be named ${newRole.name}`);
       t.equal(addedRole.description, newRole.description, `The description should be named ${newRole.description}`);
     });
@@ -396,7 +396,7 @@ test("Test create a client role - client id doesn't exist", (t) => {
     };
 
     return kca.then((client) => {
-      return t.shouldFail(client.clients.createRole(realmName, id, newRole), 'Could not find client', 'Should return an error that no client is found');
+      return t.shouldFail(client.clients.roles.create(realmName, id, newRole), 'Could not find client', 'Should return an error that no client is found');
     });
   });
 });
@@ -412,7 +412,7 @@ test('Test create a client role - a non-unique role name', (t) => {
     };
 
     return kca.then((client) => {
-      return t.shouldFail(client.clients.createRole(realmName, id, newRole), `Role ${newRole.name} already exists`, 'Error message should be returned when using a non-unique role name');
+      return t.shouldFail(client.clients.roles.create(realmName, id, newRole), `Role ${newRole.name} already exists`, 'Error message should be returned when using a non-unique role name');
     });
   });
 });
